requirements.txt
pandas>=1.5.0
numpy>=1.21.0
requests>=2.28.0
websocket-client>=1.3.0
pycryptodome>=3.15.0
python-dotenv>=0.19.0
pytz>=2022.1 